Add comprehensive curve transformations, metrics, and traits
This commit introduces a wide range of mathematical and domain-specific transformations for curves, including geometric, topological, and spectral operations. New traits for curve metrics extraction enable detailed analysis, including shape, range, trend, and risk metrics. Additionally, enhanced interpolation methods and curve construction mechanisms are added, supporting tasks such as bilinear and parametric computations. (ref. #96)
