+++ src/curves/mod.rs
@@ -5,6 +5,9 @@ mod interpolation;
 mod operations;
 mod types;
 mod visualization;
+mod curve;
+mod utils;
 
 pub use types::Point2D;
 pub use operations::arithmetic::MergeOperation;
+pub use curve::Curve;
