+++ src/curves/operations/mod.rs
@@ -1,2 +1,5 @@
 pub mod arithmetic;
 mod transformations;
+mod traits;
+
+pub use traits::CurveArithmetic;
